requirements.txt
asgiref==3.6.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.11.0
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.0.12
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-browser-reload==1.6.0
django-filter==22.1
django-tailwind==3.4.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
Pillow==9.1.1
platformdirs==2.6.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
utils==1.0.1
virtualenv==20.17.0
whitenoise==6.2.0
winshell==0.6
xendit-python==0.2.0
